introduce general resource manager
